% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAModel.r
\name{buildAModel}
\alias{buildAModel}
\title{Builds a GLM model}
\usage{
buildAModel(dataSource, predicator, numberOfKnots = NULL)
}
\arguments{
\item{dataSource}{the data frame to process.}

\item{predicator}{the independent variable.}

\item{numberOfKnots}{Number of knots - Specifying numberOfKnots=0 will force all effects to be linear which will speed up the process.}
}
\value{
text
}
\description{
Dropping the most redundant variables in a step wise fashion, the remaining variables are used to build a GLM model, then extract only the statistically significant variables and rebuild the model. Finally, generate the probability formula.
}

