% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_50.r
\name{detection_50}
\alias{detection_50}
\title{Detection of the steps - main function}
\usage{
detection_50(dataSource, predictor, cutoff, num_runs = 10)
}
\arguments{
\item{dataSource}{the data frame to process.}

\item{cutoff}{the cutoff of the detection}

\item{num_runs}{the number of runs - default = 10}

\item{predicator}{the predicator, it can be passed as a vector of predicators}
}
\value{
vector containing the mean offalse positive, accuracy, recall, precision and the f-measure of the detection based on the number of runs.
}
\description{
This can be used to detect steps using a training and testing datasets - using detection()
}

