% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eularAngles.r
\name{eularAngles}
\alias{eularAngles}
\title{Calculate the Eular angles from quaternion}
\usage{
eularAngles(timestamp, a, b, c, d, radian = TRUE)
}
\arguments{
\item{timestamp}{the timestamp of the step}

\item{a}{represents the reading from the quaternion value of quat0}

\item{b}{represents the reading from the quaternion value of quat1}

\item{c}{represents the reading from the quaternion value of quat2}

\item{d}{represents the reading from the quaternion value of quat3}

\item{radian}{by default is TRUE, if FALSE the returned calculations would be in degrees instead.}
}
\value{
a list contains the following data (timestamp, roll1, roll2, pitch, yaw1, yaw2)
}
\description{
Calculate the eular angles from quaternion and return (timestamp, roll1, roll2, pitch, yaw1, yaw2)
}

