% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractStepCycle.r
\name{extractStepCycle}
\alias{extractStepCycle}
\title{Extract the step cycle}
\usage{
extractStepCycle(pathOfTheFolder, sep = ";")
}
\arguments{
\item{pathOfTheFolder}{path to the source of the raw data.}

\item{sep}{the field separator character.}
}
\value{
a data frame of the steps cycle.
}
\description{
Extract the step cycle of a session.
}

