% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalDataSet.r
\docType{data}
\name{finalDataSet}
\alias{finalDataSet}
\title{2029 steps of the testing data}
\format{A data frame with 2029 rows and 120 variables}
\usage{
data(finalDataSet)
}
\description{
A dataset containing the steps of the experiments conducted earlier for the study where step is (0, 1, 2) as (up, down and flat) respectively.
}
\keyword{datasets}

