% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRMSE.r
\name{findRMSEFoundation}
\alias{findRMSEFoundation}
\title{Builds a predictive model and returns the RMSE}
\usage{
findRMSEFoundation(dataSource, train, test, predictor)
}
\arguments{
\item{dataSource}{the data frame to process.}

\item{train}{the training data frame.}

\item{test}{the testing data frame.}

\item{predictor}{the predictor to build the model with.}
}
\value{
matrix
}
\description{
Builds a predictive model and return the Root-Mean-Square Error (the difference between the actual and the predicted values).
}

