% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDataSet.r
\name{generateDataSet}
\alias{generateDataSet}
\title{Generate dataset containing the metrics}
\usage{
generateDataSet(pathOfTheFolder, features, step, sep = ";")
}
\arguments{
\item{pathOfTheFolder}{path to the source of the raw data.}

\item{step}{indicates the step, we used {up, down and flat} as {0,1,2} respectively.}

\item{sep}{the field separator character - default ";".}

\item{feature}{Which feature to calculate, should be passed as a string or a vector of strings.}
}
\value{
a data frame
}
\description{
Detecting steps using a training and testing datasets
}

