% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDataSetsLocally.r
\name{generateDataSetsLocally}
\alias{generateDataSetsLocally}
\title{Generate dataset containing the steps data - from the local files of the package}
\usage{
generateDataSetsLocally(folder, step, sep = ",")
}
\arguments{
\item{folder}{path to the source of the raw data.}

\item{step}{indicates the step, we used {up, down and flat} as {0,1,2} respectively.}

\item{sep}{the field separator character - default ";".}

\item{feature}{Which feature to calculate, should be passed as a string or a vector of strings.}
}
\value{
a data frame
}
\description{
Generate dataset containing the steps data
}

