% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateFullFeatures.r
\name{generateFullFeatures}
\alias{generateFullFeatures}
\title{Generate a dataset with all the features}
\usage{
generateFullFeatures(steps, stepsCycle, feature, step)
}
\arguments{
\item{steps}{The data frame of the steps.}

\item{stepsCycle}{The data frame of the steps' cycle.}

\item{step}{indicates the step, we used {up, down and flat} as {0,1,2} respectively.}
}
\value{
data frame containing the step segments and with a prefix of the feature in the column names.
}
\description{
Merge session files into a data frame and calculate each feature by the step segment of that session.
}

