% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalClustering_Spearman.r
\name{hierarchicalClustering_Spearman}
\alias{hierarchicalClustering_Spearman}
\title{Plot the hierarchical Clustering}
\usage{
hierarchicalClustering_Spearman(dataSource, threshold = 0.5)
}
\arguments{
\item{dataSource}{the data frame to process.}

\item{threshold}{the threshold of the line of the highly correlated columns - default = 0.5}
}
\value{
a plot containg a hierarchical presentation of the highly correlated values
}
\description{
Plot the hierarchical Clustering using sperman correlation of the highly correlated columns.
}

