% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeAllRawData.r
\name{mergeAllRawData}
\alias{mergeAllRawData}
\title{Merge all the raw data into one dataframe}
\usage{
mergeAllRawData(pathOfTheFolder, step, radian = TRUE, sep = ";")
}
\arguments{
\item{pathOfTheFolder}{path to the source of the raw data.}

\item{step}{indicates the step, we used {up, down and flat} as {0,1,2} respectively.}

\item{radian}{by default is TRUE, if FALSE the returned calculations would be in degrees instead.}

\item{sep}{the field separator character - default ";".}
}
\value{
a data frame of all the data merged.
}
\description{
Merge all the raw data into one dataframe and return a data frame of all the data merged.
}

