% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAndApplyFeature.r
\name{mergeAndApplyFeature}
\alias{mergeAndApplyFeature}
\title{Merge session files into one data frame}
\usage{
mergeAndApplyFeature(pathOfTheFolder, feature)
}
\arguments{
\item{pathOfTheFolder}{The full path of the session.}

\item{feature}{Which feature to calculate, should be passed as a string}
}
\value{
data frame containing the step segments and with a prefix of the feature in the column names.
}
\description{
Merge session files into a data frame and calculate each feature by the step segment of that session.
}

