% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeAndApplyFeatureDataFrameLocally.r
\name{mergeAndApplyFeatureDataFrameLocally}
\alias{mergeAndApplyFeatureDataFrameLocally}
\title{Merge session files into one data frame - on the raw files shipped with the package}
\usage{
mergeAndApplyFeatureDataFrameLocally(dataSource, stepsCycle, feature)
}
\arguments{
\item{dataSource}{the data frame to process.}

\item{stepsCycle}{the step cycle data frame}

\item{feature}{Which feature to calculate, should be passed as a string}
}
\value{
data frame containing the step segments and with a prefix of the feature in the column names.
}
\description{
Merge session files into a data frame and calculate each feature by the step segment of that session.
}

