% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.r
\name{R2}
\alias{R2}
\title{Calculate the explained deviance of a model}
\usage{
R2(model)
}
\arguments{
\item{model}{the model to extract the explained deviance from.}
}
\value{
text
}
\description{
Calculate the fitness of a model using the following forumla: R-squared = 1 - (estimated variance)/(observed variance)
}

