% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCorrelatedFeatures.r
\name{removeCorrelatedFeatures}
\alias{removeCorrelatedFeatures}
\title{Remove the highly correlated features}
\usage{
removeCorrelatedFeatures(dataSource, predicator = "step", threshold = 0.9,
  numberOfKnots = NULL)
}
\arguments{
\item{dataSource}{the data frame to process.}

\item{predicator}{the independent variable - default = step.}

\item{threshold}{the adjusted R^2 cutoff for redundancy - default = 0.9}

\item{numberOfKnots}{Number of knots - Specifying numberOfKnots=0 will force all effects to be linear which will speed up the process.}
}
\value{
text
}
\description{
Dropping the most redundant variables in a step wise fashion, the remaining variables along with the predicator are returned in a forumla.
}

