% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalNumberOfMetricsUsed.r
\name{totalNumberOfMetricsUsed}
\alias{totalNumberOfMetricsUsed}
\title{Display the number of metrics}
\usage{
totalNumberOfMetricsUsed(model)
}
\arguments{
\item{model}{The model equation}
}
\value{
text
}
\description{
Display the number of metrics used in the probability equation
}

