% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{buildAModel}
\alias{buildAModel}
\title{Builds a GLM model}
\usage{
buildAModel(dataSource, predicator)
}
\arguments{
\item{dataSource}{the data frame to process.}

\item{predicator}{the independent variable.}
}
\value{
text
}
\description{
Dropping the most redundant variables in a step wise fashion, the remaining variables are used to build a GLM model, then extract only the statistically significant variables and rebuild the model. Finally, generate the probability formula.
}

