% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAModel.r
\name{buildAModel}
\alias{buildAModel}
\title{Builds a GLM model}
\usage{
buildAModel(dataSource, predicator, forumla, threshold = 0.05)
}
\arguments{
\item{dataSource}{the data frame to process.}

\item{predicator}{the independent variable.}

\item{forumla}{list of metrics concatenated with a plus.}

\item{threshold}{the threshold of the significant variables}
}
\value{
text
}
\description{
Dropping the most redundant variables in a step wise fashion, the remaining variables are used to build a GLM model, then extract only the statistically significant variables and rebuild the model. Finally, generate the probability formula.
}

