% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eularAngles.r
\name{eularAngles}
\alias{eularAngles}
\title{Calculate the Eular angles from quaternion}
\usage{
eularAngles(timestamp, a, b, c, d, radian = TRUE)
}
\arguments{
\item{timestamp}{the timestamp of the step}

\item{radian}{by default is TRUE, if FALSE the returned calculations would be in degrees instead.}

\item{(a, b, c, d)}{represents the reading from the quaternion (quat0,quat1,quat2,quat3) respectively.}
}
\value{
a list contains the following data (timestamp, roll1, roll2, pitch, yaw1, yaw2)
}
\description{
Calculate the eular angles from quaternion and return (timestamp, roll1, roll2, pitch, yaw1, yaw2)
}

