% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.r
\name{r2}
\alias{r2}
\title{Calculate the explained deviance of a model}
\usage{
r2(model)
}
\arguments{
\item{model}{the model to extract the explained deviance from.}
}
\value{
text
}
\description{
Calculate the fitness of a model using the following forumla: R-squared = 1 - (estimated variance)/(observed variance)
}

